check use of lib
